Refactoring system to use one hand per context

+++ src/graphics/VolumeControl.cpp
@@ -6,10 +6,10 @@
 #include <iostream>
 
 VolumeControl::VolumeControl(float radius, float width) :
-m_partialDisk(new PartialDisk()),
-m_activePartialDisk(new PartialDisk()),
-m_needle(new RectanglePrim()),
-m_maxOpacity(0.8f)
+  m_partialDisk(new PartialDisk()),
+  m_activePartialDisk(new PartialDisk()),
+  m_needle(new RectanglePrim()),
+  m_maxOpacity(0.8f)
 {
   m_partialDisk->SetDiffuseColor(Color(0.7f, 0.7f, 0.7f,m_maxOpacity));
   m_partialDisk->SetAmbientFactor(0.9f);
@@ -44,7 +44,7 @@ void VolumeControl::InitChildren() {
 }
 
 float VolumeControl::Volume() {
-  return volumeFromAngle(m_activePartialDisk->EndAngle());
+  return volumeFromAngle((float)m_activePartialDisk->EndAngle());
 }
 
 void VolumeControl::SetVolume(float volume) {
@@ -55,7 +55,7 @@ void VolumeControl::SetVolume(float volume) {
 
 void VolumeControl::NudgeVolume(float dVolume) {
   std::cout << "dVolume: " << dVolume << std::endl;
-  float angle = m_activePartialDisk->EndAngle() + VolumeControl::dAngleFromVolume(dVolume);
+  float angle = (float)m_activePartialDisk->EndAngle() + VolumeControl::dAngleFromVolume(dVolume);
   m_activePartialDisk->SetEndAngle(angle);
   std::cout << "New Volume: " << volumeFromAngle(angle) << std::endl;
 }
@@ -75,13 +75,13 @@ void VolumeControl::SetOpacity(float opacity) {
 }
 
 float VolumeControl::volumeFromAngle(float angle) {
-  return (angle + 5*M_PI/4)/(3*M_PI/2);
+  return (float)((angle + 5*M_PI/4)/(3*M_PI/2));
 }
 
 float VolumeControl::angleFromVolume(float volume) {
-  return ((3*M_PI)/2)*volume - (5*M_PI)/4;
+  return (float)(((3 * M_PI) / 2)*volume - (5 * M_PI) / 4);
 }
 
 float VolumeControl::dAngleFromVolume(float dVolume) {
-  return ((3*M_PI)/2)*dVolume;
+  return (float) (((3 * M_PI) / 2)*dVolume);
 }
